refactor: use saparate component for network filter
