+++ app/markets/components/markets.tsx
@@ -19,6 +19,7 @@ import {
   infoToKey,
 } from '@/utils/tokens';
 
+import NetworkFilter from './NetworkFilter';
 import MarketsTable from './marketsTable';
 import { SupplyModal } from './supplyModal';
 
@@ -238,49 +239,7 @@ export default function HomePage() {
           {/* left section: asset filters */}
           <div className="flex flex-col gap-2 lg:flex-row">
             {/* network filter */}
-            <Select
-              label="Network"
-              selectionMode="single"
-              placeholder="All networks"
-              onChange={(e) => {
-                if (!e.target.value) setSelectedNetwork(null);
-                const newId = Number(e.target.value);
-                if (newId && isSupportedChain(newId)) {
-                  setSelectedNetwork(newId);
-                }
-              }}
-              classNames={{
-                trigger: 'bg-secondary rounded-sm min-w-32',
-                popoverContent: 'bg-secondary rounded-sm',
-              }}
-              items={networks}
-              renderValue={(items) => {
-                return (
-                  <div className="flex-scroll flex gap-1">
-                    {items.map((item) => {
-                      const networkImg = getNetworkImg(Number(item.key));
-                      return networkImg ? (
-                        <Image src={networkImg} alt="icon" height="18" />
-                      ) : (
-                        item.textValue
-                      );
-                    })}
-                  </div>
-                );
-              }}
-            >
-              {networks.map((network) => {
-                // key = `0x5A98FcBEA516Cf06857215779Fd812CA3beF1B32-1|0x5A98FcBEA516Cf06857215779Fd812CA3beF1B32-42`
-                return (
-                  <SelectItem key={network.network} textValue={network.name}>
-                    <div className="flex items-center justify-between">
-                      <p>{network.name}</p>
-                      <Image className="ml-auto" src={network.logo} alt="icon" height="18" />
-                    </div>
-                  </SelectItem>
-                );
-              })}
-            </Select>
+            <NetworkFilter selectedNetwork={selectedNetwork} setSelectedNetwork={setSelectedNetwork} />
 
             <Select
               label="Loan Asset"
